dash/requirements.txt
dash 
dash-daq 
dash-bootstrap-components
cassandra-driver
pandas
